+++ upload/catalog/controller/extension/payment/payzen.php
@@ -1,25 +1,11 @@
 <?php
 /**
- * PayZen V2-Payment Module version 4.0.0 for OpenCart 3.x. Support contact : support@payzen.eu.
+ * Copyright © Lyra Network.
+ * This file is part of PayZen plugin for OpenCart See COPYING.md for license details.
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- * @author    Lyra Network (http://www.lyra-network.com/)
- * @copyright 2014-2018 Lyra Network and contributors
- * @license   http://www.gnu.org/licenses/gpl.html  GNU General Public License (GPL v3)
- * @category  payment
- * @package   payzen
+ * @author     Lyra Network <https://www.lyra.com>
+ * @copyright  Lyra Network
+ * @license    http://www.gnu.org/licenses/gpl.html GNU General Public License (GPL v3)
  */
 
 class ControllerExtensionPaymentPayzen extends Controller
@@ -40,7 +26,7 @@ class ControllerExtensionPaymentPayzen extends Controller
         require_once(DIR_SYSTEM . 'library/payzen/tools.php');
         $this->plugin_features = PayzenTools::$plugin_features;
 
-        // init class logger
+        // Init class logger.
         $this->logger = new Log(date('Y_m') . '_payzen.log');
     }
 
@@ -48,7 +34,7 @@ class ControllerExtensionPaymentPayzen extends Controller
     {
         $payzenRequest = $this->getPayzenRequest();
 
-        // define template variables
+        // Define template variables.
         $data = array(
             'payzen_form_fields' => $payzenRequest->getRequestHtmlFields(),
             'payzen_form_action' => $payzenRequest->get('platform_url'),
@@ -57,7 +43,7 @@ class ControllerExtensionPaymentPayzen extends Controller
             'button_confirm' => $this->language->get('button_confirm')
         );
 
-        $this->writeLog('Parameters to send to payment gateway : ' . print_r($payzenRequest->getRequestFieldsArray(), true), 'DEBUG');
+        $this->writeLog('Data to be sent to payment gateway: ' . print_r($payzenRequest->getRequestFieldsArray(true /* To hide sensitive data. */), true), 'DEBUG');
 
         return $this->load->view('extension/payment/payzen', $data);
     }
@@ -71,7 +57,7 @@ class ControllerExtensionPaymentPayzen extends Controller
         $orderInfo = $this->model_checkout_order->getOrder($this->session->data['order_id']);
 
         if ($orderInfo) {
-            // reset order to pending status if it isn't
+            // Reset order to pending status if it isn't.
             if ($this->isOrderPending($orderInfo)) {
                 $this->db->query(
                     "UPDATE `" . DB_PREFIX . "order` SET order_status_id = '0', date_modified = NOW() WHERE order_id = '" .
@@ -79,32 +65,32 @@ class ControllerExtensionPaymentPayzen extends Controller
                 );
             }
 
-            // update method title to be displayed in backend
+            // Update method title to be displayed in backend.
             $this->load->model('extension/payment/' . $this->name);
             $this->{'model_extension_payment_' . $this->name}->updateMethodTitle($orderInfo['order_id']);
 
             $info = array();
 
-            // get used currency
+            // Get used currency.
             $currency = PayzenApi::findCurrencyByAlphaCode($orderInfo['currency_code']);
             $info['currency'] = $currency->getNum();
 
-            // get store language
+            // Get store language.
             $currentLang = strtolower(substr($orderInfo['language_code'], 0, 2));
             $configLang = $this->config->get($this->prefix . $this->name . '_language');
             $info['language'] = PayzenApi::isSupportedLanguage($currentLang) ? $currentLang : $configLang ;
 
-            // misc parameters
+            // Misc parameters.
             $info['order_id'] = $orderInfo['order_id'];
 
             $amount = $this->currency->format($orderInfo['total'], $orderInfo['currency_code'], $orderInfo['currency_value'], false);
             $info['amount'] = $currency->convertAmountToInteger($amount);
 
-            $info['contrib'] = 'OpenCart3.x_4.0.0/' . VERSION . '/' . PHP_VERSION;
+            $info['contrib'] = 'OpenCart_3.x_4.1.0/' . VERSION . '/' . PHP_VERSION;
 
             $info['order_info'] = 'session_id=' . $this->session->getId();
 
-            // customer info
+            // Customer info.
             $info['cust_id'] = $orderInfo['customer_id'];
             $info['cust_email'] = $orderInfo['email'];
             $info['cust_phone'] = $orderInfo['telephone'];
@@ -117,7 +103,7 @@ class ControllerExtensionPaymentPayzen extends Controller
             $info['cust_city'] = $orderInfo['payment_city'];
             $info['cust_state'] = $orderInfo['payment_zone'];
 
-            // customer shipping address
+            // Customer shipping address.
             $info['ship_to_first_name'] = $orderInfo['shipping_firstname'] ;
             $info['ship_to_last_name'] = $orderInfo['shipping_lastname'];
             $info['ship_to_street'] = $orderInfo['shipping_address_1'];
@@ -127,19 +113,19 @@ class ControllerExtensionPaymentPayzen extends Controller
             $info['ship_to_country'] = $orderInfo['shipping_iso_code_2'];
             $info['ship_to_zip'] = $orderInfo['shipping_postcode'];
 
-            // activate 3DS ?
+            // Activate 3DS?
             $threedsMpi = null;
             $threedsMinAmount = $this->config->get($this->prefix . $this->name . '_3ds_min_amount');
-            if ($threedsMinAmount != '' && $orderInfo['total'] < $threedsMinAmount) {
+            if ($threedsMinAmount && $orderInfo['total'] < $threedsMinAmount) {
                 $threedsMpi = '2';
             }
 
             $info['threeds_mpi'] = $threedsMpi;
 
-            // return URL
+            // Return URL.
             $info['url_return'] = $this->url->link('extension/payment/' . $this->name . '/process');
 
-            // admin configuration parameters
+            // Admin configuration parameters.
             $configParams = array(
                 'site_id', 'key_test', 'key_prod', 'ctx_mode', 'platform_url', 'available_languages',
                 'capture_delay', 'validation_mode', 'payment_cards', 'redirect_enabled',
@@ -164,7 +150,6 @@ class ControllerExtensionPaymentPayzen extends Controller
         $data = $this->clean($this->request->post);
 
         $this->writeLog('Begin server response processing.');
-        $this->writeLog('Parameters received from payment gateway : ' . print_r($data, true), 'DEBUG');
 
         require_once(DIR_SYSTEM . 'library/payzen/response.php');
         $payzenResponse = new PayzenResponse(
@@ -178,42 +163,41 @@ class ControllerExtensionPaymentPayzen extends Controller
         if (! $payzenResponse->isAuthentified()) {
             $ip = $_SERVER['REMOTE_ADDR'];
 
-            $this->writeLog("{$ip} tries to access extension/payment/payzen/callback page without valid signature with parameters: " . print_r($data, true), 'ERROR');
+            $this->writeLog("{$ip} tries to access payzen/callback page without valid signature with parameters: " . print_r($data, true), 'ERROR');
             $this->writeLog('Signature algorithm selected in module settings must be the same as one selected in PayZen Back Office.', 'ERROR');
-            die($payzenResponse->getOutputForPlatform('auth_fail'));
+            die($payzenResponse->getOutputForGateway('auth_fail'));
         }
 
         $orderId = $payzenResponse->get('order_id');
         $orderInfo = $this->model_checkout_order->getOrder($orderId);
         if (empty($orderInfo)) {
-            $this->writeLog('Order with ID#' . $orderId . ' was not found.', 'ERROR');
+            $this->writeLog("Order with ID #$orderId was not found.", 'ERROR');
             $this->writeLog('End server response processing.');
-            die($payzenResponse->getOutputForPlatform('order_not_found'));
+            die($payzenResponse->getOutputForGateway('order_not_found'));
         }
 
         if ($this->isOrderPending($orderInfo)) {
-            $this->writeLog('First payment notification. Let\'s change order status.');
+            $this->writeLog("First payment notification for order #$orderId. Let's change order status.");
 
             if ($payzenResponse->isAcceptedPayment()) {
-                $this->writeLog('Payment accepted. New order status is ' . $this->config->get($this->prefix . $this->name . '_order_status_success') . '.');
+                $this->writeLog('Payment for order #' . $orderId . ' has been confirmed by notification URL. New order status is ' . $this->config->get($this->prefix . $this->name . '_order_status_success') . '.');
                 $this->model_checkout_order->addOrderHistory(
                     $orderId,
                     $this->config->get($this->prefix . $this->name . '_order_status_success'),
-                    $payzenResponse->getMessage(),
+                    $payzenResponse->getCompleteMessage(),
                     true
                 );
 
-                $this->writeLog('Clear cart and session vars.');
+                $this->writeLog("Clear cart and session vars to process order #$orderId.");
 
-                // destroy current session and restore the session in which the payment was initiated
-                $parts = explode('&', $payzenResponse->get('order_info'));
+                // Destroy current session and restore the session in which the payment was initiated.
                 $sessionId = substr($payzenResponse->get('order_info'), strlen('session_id='));
 
                 $this->session->__destroy();
                 $this->session->start($sessionId);
                 $this->registry->set('session', $this->session);
 
-                // clear cart and delete used session vars
+                // Clear cart and delete used session vars.
                 if (isset($this->session->data['order_id'])) {
                     $customer = new Cart\Customer($this->registry);
                     $this->registry->set('customer', $customer);
@@ -234,7 +218,7 @@ class ControllerExtensionPaymentPayzen extends Controller
                 }
 
                 $this->writeLog('End server response processing.');
-                die($payzenResponse->getOutputForPlatform('payment_ok'));
+                die($payzenResponse->getOutputForGateway('payment_ok'));
             } else {
                 $orderStatus = $payzenResponse->isCancelledPayment() ?
                     $this->config->get($this->prefix . $this->name . '_order_status_canceled') :
@@ -244,33 +228,33 @@ class ControllerExtensionPaymentPayzen extends Controller
                     ($this->config->get($this->prefix . $this->name . '_notify_canceled') === '1') :
                     ($this->config->get($this->prefix . $this->name . '_notify_failed') === '1');
 
-                $this->writeLog('Payment not accepted. New order status is ' . $orderStatus . '.');
+                $this->writeLog("Payment for order #$orderId not accepted. New order status is $orderStatus.");
 
                 $this->model_checkout_order->addOrderHistory(
                     $orderId,
                     $orderStatus,
-                    $payzenResponse->getMessage(),
+                    $payzenResponse->getCompleteMessage(),
                     $notify
                 );
 
                 $this->writeLog('End server response processing.');
-                die($payzenResponse->getOutputForPlatform('payment_ko'));
+                die($payzenResponse->getOutputForGateway('payment_ko'));
             }
         } else {
-            // order already processed
+            // Order already processed.
 
             if ($this->isOrderSuccess($orderInfo) && $payzenResponse->isAcceptedPayment()) {
-                $this->writeLog('Payment notification confirmed. Payment accepted.');
+                $this->writeLog("Payment notification confirmed. Payment accepted for order #$orderId.");
                 $this->writeLog('End server response processing.');
-                die($payzenResponse->getOutputForPlatform('payment_ok_already_done'));
+                die($payzenResponse->getOutputForGateway('payment_ok_already_done'));
             } elseif (($this->isOrderFailed($orderInfo) || $this->isOrderCanceled($orderInfo)) && ! $payzenResponse->isAcceptedPayment()) {
-                $this->writeLog('Payment notification confirmed. Payment error or cancelation.');
+                $this->writeLog("Payment notification confirmed. Payment failed or cancelled for order #$orderId.");
                 $this->writeLog('End server response processing.');
-                die($payzenResponse->getOutputForPlatform('payment_ko_already_done'));
+                die($payzenResponse->getOutputForGateway('payment_ko_already_done'));
             } else {
-                $this->writeLog('Error! Invalid payment code received for an already processed order. Order status is ' . $orderInfo['order_status_id'] . '. Payment result is ' . $payzenResponse->get('code'), 'ERROR');
+                $this->writeLog('Error! Invalid payment code received for an already processed order #' . $orderId . '. Order status is ' . $orderInfo['order_status_id'] . '. Payment result is ' . $payzenResponse->get('code'), 'ERROR');
                 $this->writeLog('End server response processing.');
-                die($payzenResponse->getOutputForPlatform('payment_ko_on_order_ok'));
+                die($payzenResponse->getOutputForGateway('payment_ko_on_order_ok'));
             }
         }
     }
@@ -296,12 +280,12 @@ class ControllerExtensionPaymentPayzen extends Controller
         );
 
         if (! $payzenResponse->isAuthentified()) {
-            // fatal error: clear cart
+            // Fatal error: clear cart.
             $this->cart->clear();
 
             $ip = $_SERVER['REMOTE_ADDR'];
 
-            $this->writeLog("{$ip} tries to access extension/payment/payzen/process page without valid signature with parameters: " . print_r($data, true), 'ERROR');
+            $this->writeLog("{$ip} tries to access payzen/process page without valid signature with parameters: " . print_r($data, true), 'ERROR');
             $this->writeLog('Signature algorithm selected in module settings must be the same as one selected in PayZen Back Office.', 'ERROR');
 
             $this->writeLog('End return user response processing.');
@@ -312,16 +296,16 @@ class ControllerExtensionPaymentPayzen extends Controller
         $orderId = $payzenResponse->get('order_id');
         $orderInfo = $this->model_checkout_order->getOrder($orderId);
         if (empty($orderInfo)) {
-            // fatal error: clear cart
+            // Fatal error: clear cart.
             $this->cart->clear();
 
-            $this->writeLog('Order with ID#' . $orderId . ' was not found. Clear cart and display error message.', 'ERROR');
+            $this->writeLog("Order with ID #$orderId was not found.", 'ERROR');
             $this->writeLog('End return user response processing.');
             $this->renderResponse(true);
             return;
         }
 
-        // load payment currency
+        // Load payment currency.
         $currency = PayzenApi::findCurrencyByNumCode($payzenResponse->get('currency'));
         $data = array(
             'trans_id' => $payzenResponse->get('trans_id'),
@@ -329,23 +313,23 @@ class ControllerExtensionPaymentPayzen extends Controller
         );
 
         if ($this->isOrderPending($orderInfo)) {
-            $this->writeLog('First payment notification by user return, server URL notification does not work. Let\'s change order status and show warning to merchant.', 'WARN');
+            $this->writeLog("First payment notification by user return for order #$orderId, server URL notification does not work. Let's change order status and show warning to merchant.", 'WARN');
 
-            // is it TEST mode ?
-            $testMode = ($this->config->get($this->prefix . $this->name . '_ctx_mode') == 'TEST');
+            // Is it TEST mode?
+            $testMode = ($this->config->get($this->prefix . $this->name . '_ctx_mode') === 'TEST');
 
             if ($payzenResponse->isAcceptedPayment()) {
-                $this->writeLog('Payment accepted. New order status is ' . $this->config->get($this->prefix . $this->name . '_order_status_success') . '.');
+                $this->writeLog("Payment for order #$orderId accepted. New order status is " . $this->config->get($this->prefix . $this->name . '_order_status_success') . '.');
                 $this->model_checkout_order->addOrderHistory(
                     $orderId,
                     $this->config->get($this->prefix . $this->name . '_order_status_success'),
-                    $payzenResponse->getLogString(),
+                    $payzenResponse->getCompleteMessage(),
                     true
                 );
 
-                $this->writeLog('Clear cart and session vars.');
+                $this->writeLog("Clear cart and session vars to process order #$orderId.");
 
-                // clear cart and delete used session vars
+                // Clear cart and delete used session vars.
                 if (isset($this->session->data['order_id'])) {
                     $this->cart->clear();
 
@@ -369,15 +353,15 @@ class ControllerExtensionPaymentPayzen extends Controller
                     $this->config->get($this->prefix . $this->name . '_order_status_failed');
 
                 $notify = $payzenResponse->isCancelledPayment() ?
-                    ($this->config->get($this->prefix . $this->name . '_notify_canceled') == '1') :
-                    ($this->config->get($this->prefix . $this->name . '_notify_failed') == '1');
+                    ($this->config->get($this->prefix . $this->name . '_notify_canceled') === '1') :
+                    ($this->config->get($this->prefix . $this->name . '_notify_failed') === '1');
 
-                $this->writeLog('Payment not accepted. New order status is ' . $orderStatus . '.');
+                    $this->writeLog("Payment for order #$orderId not accepted. New order status is $orderStatus.");
 
                 $this->model_checkout_order->addOrderHistory(
                     $orderId,
                     $orderStatus,
-                    $payzenResponse->getLogString(),
+                    $payzenResponse->getCompleteMessage(),
                     $notify
                 );
 
@@ -389,24 +373,24 @@ class ControllerExtensionPaymentPayzen extends Controller
                 }
             }
         } else {
-            // Order already processed
+            // Order already processed.
             if ($this->isOrderSuccess($orderInfo) && $payzenResponse->isAcceptedPayment()) {
-                $this->writeLog('Payment notification confirmed. Payment accepted.');
+                $this->writeLog("Payment notification confirmed. Payment accepted for order #$orderId.");
                 $this->writeLog('End user return response processing.');
                 $this->renderResponse(false, false, $data);
             } elseif ($this->isOrderCanceled($orderInfo) && $payzenResponse->isCancelledPayment()) {
-                $this->writeLog('Payment notification confirmed. Payment canceled.');
+                $this->writeLog("Payment notification confirmed. Payment cancelled for order #$orderId.");
                 $this->writeLog('End user return response processing.');
                 $this->response->redirect($this->url->link('checkout/checkout'));
             } elseif ($this->isOrderFailed($orderInfo) && ! $payzenResponse->isAcceptedPayment() && ! $payzenResponse->isCancelledPayment()) {
-                $this->writeLog('Payment notification reconfirmed. Payment failed.');
+                $this->writeLog("Payment notification confirmed. Payment failed for order #$orderId.");
                 $this->writeLog('End user return response processing.');
                 $this->renderResponse($this->language->get('text_payment_payzen_payment_error'));
             } else {
-                // fatal error: clear cart
+                // Fatal error: clear cart.
                 $this->cart->clear();
 
-                $this->writeLog('Error! Invalid payment code received for an already processed order. Order status is ' . $orderInfo['order_status_id'] . '. Payment result is ' . $payzenResponse->get('code'), 'ERROR');
+                $this->writeLog('Error! Invalid payment code received for an already processed order #' . $orderId . '. Order status is ' . $orderInfo['order_status_id'] . '. Payment result is ' . $payzenResponse->get('code'), 'ERROR');
                 $this->writeLog('End user return response processing.');
                 $this->renderResponse(true);
             }
@@ -415,16 +399,16 @@ class ControllerExtensionPaymentPayzen extends Controller
 
     private function renderResponse($errorMessage = false, $urlCheckWarn = false, $data = array())
     {
-        // set misc data (when payment is successful)
+        // Set misc data (when payment is successful).
         $data['payzen_data'] = array();
         foreach ($data as $name => $value) {
-            if ($value != null) {
+            if ($value) {
                 $key = $this->language->get('entry_payment_payzen_' . $name);
                 $data['payzen_data'][$key] = $value ;
             }
         }
 
-        // set check URL warning
+        // Set check URL warning.
         if ($urlCheckWarn) {
             if ($this->config->get('config_maintenance')) {
                 $data['text_payment_payzen_url_check'] = $this->language->get('text_payment_payzen_maint_warn');
@@ -435,10 +419,9 @@ class ControllerExtensionPaymentPayzen extends Controller
             }
         }
 
-        // set going to prod information
-        if ($this->config->get($this->prefix . $this->name . '_ctx_mode') == 'TEST' && $this->plugin_features['prodfaq']) {
+        // Set going to prod information.
+        if ($this->config->get($this->prefix . $this->name . '_ctx_mode') === 'TEST' && $this->plugin_features['prodfaq']) {
             $data['text_payment_payzen_pass_to_prod'] = $this->language->get('text_payment_payzen_pass_to_prod_info');
-            $data['text_payment_payzen_pass_to_prod'] .= '<a href="https://payzen.io/en-EN/faq/how-to-switch-my-shop-to-production-mode.html" target="_blank">https://payzen.io/en-EN/faq/how-to-switch-my-shop-to-production-mode.html</a>';
         }
 
         if ($errorMessage) {
@@ -492,22 +475,22 @@ class ControllerExtensionPaymentPayzen extends Controller
 
     private function isOrderPending($orderInfo)
     {
-        return $orderInfo['order_status_id'] == 0;
+        return $orderInfo['order_status_id'] === '0';
     }
 
     private function isOrderSuccess($orderInfo)
     {
-        return $orderInfo['order_status_id'] == $this->config->get($this->prefix . $this->name . '_order_status_success');
+        return $orderInfo['order_status_id'] === $this->config->get($this->prefix . $this->name . '_order_status_success');
     }
 
     private function isOrderFailed($orderInfo)
     {
-        return $orderInfo['order_status_id'] == $this->config->get($this->prefix . $this->name . '_order_status_failed');
+        return $orderInfo['order_status_id'] === $this->config->get($this->prefix . $this->name . '_order_status_failed');
     }
 
     private function isOrderCanceled($orderInfo)
     {
-        return $orderInfo['order_status_id'] == $this->config->get($this->prefix . $this->name . '_order_status_canceled');
+        return $orderInfo['order_status_id'] === $this->config->get($this->prefix . $this->name . '_order_status_canceled');
     }
 
     /**