+++ upload/catalog/language/de-de/extension/payment/payzen.php
@@ -1,41 +1,27 @@
 <?php
 /**
- * PayZen V2-Payment Module version 4.0.0 for OpenCart 3.x. Support contact : support@payzen.eu.
+ * Copyright © Lyra Network.
+ * This file is part of PayZen plugin for OpenCart See COPYING.md for license details.
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- * @author    Lyra Network (http://www.lyra-network.com/)
- * @copyright 2014-2018 Lyra Network and contributors
- * @license   http://www.gnu.org/licenses/gpl.html  GNU General Public License (GPL v3)
- * @category  payment
- * @package   payzen
+ * @author     Lyra Network <https://www.lyra.com>
+ * @copyright  Lyra Network
+ * @license    http://www.gnu.org/licenses/gpl.html GNU General Public License (GPL v3)
  */
 
-// Text
+// Text.
 $_['text_payment_payzen_title'] = 'Zahlung mit EC-/Kreditkarte';
-$_['text_payment_payzen_multi_title'] = 'Zahlung mit EC-/Kreditkarte in %s mal';
+$_['text_payment_payzen_multi_title'] = 'Ratenzahlung mit EC-/Kreditkarte';
 
-$_['heading_title'] = 'Your order has been processed !';
-$_['heading_error_title'] = 'Your payment has been refused !';
+$_['heading_title'] = 'Your order has been processed!';
+$_['heading_error_title'] = 'Ihre Zahlung wurde abgelehnt!';
 
 $_['text_payment_payzen_maint_warn'] = 'Dieser Shop befindet sich im Wartungsmodus. Es kann keine automatische Benachrichtigung erstellt werden.';
-$_['text_payment_payzen_url_check_warn'] = 'Es konnte keine automatische Benachrichtigung erstellt werden. Bitte pr&uuml;fen Sie, ob die Benachrichtigung-URL in Ihrem Shopsystem korrekt eingerichtet ist.';
-$_['text_payment_payzen_url_check_details'] = 'N&auml;here Informationen zu diesem Problem entnehmen Sie bitte der Moduldokumentation:<br />&nbsp;&nbsp;&nbsp;- Kapitel &laquo; Bitte vor dem Weiterlesen aufmerksam lesen &raquo;<br />&nbsp;&nbsp;&nbsp;- Kapitel &laquo; Benachrichtigung-URL Einstellungen &raquo;';
-$_['text_payment_payzen_pass_to_prod_info'] = '<p><u>AUFNAHME DES PRODUKTIVBETRIEBS</u></p>Wenn Sie wissen möchten, wie Sie Ihren Shop in den Produktivbetrieb schalten können, rufen Sie folgende URL auf: ';
+$_['text_payment_payzen_url_check_warn'] = 'Es konnte keine automatische Benachrichtigung erstellt werden. Bitte prüfen Sie, ob die Benachrichtigung-URL in Ihr PayZen Back Office korrekt eingerichtet ist.';
+$_['text_payment_payzen_url_check_details'] = 'Nähere Informationen zu diesem Problem entnehmen Sie bitte der Moduldokumentations:<br /> - Kapitel « Bitte vor dem Weiterlesen aufmerksam lesen »<br /> - Kapitel « Benachrichtigung URL-Einstellungen »';
+$_['text_payment_payzen_pass_to_prod_info'] = '<p><u>UMSTELLUNG AUF PRODUKTIONSUMFELD</u></p>Sie möchten wissen, wie Sie auf Produktionsumfeld umstellen können, bitte lesen Sie die Kapitel « Weiter zur Testphase » und « Verschieben des Shops in den Produktionsumfeld » in der Dokumentation des Moduls.';
 
-$_['text_payment_payzen_fatal_error'] = '<p>Ein Fehler ist w&auml;hrend dem Zahlungsvorgang unterlaufen. Please, direct any questions you have to the <a href="%s">store owner</a>.</p>';
-$_['text_payment_payzen_payment_error'] = '<p>Ihre Zahlung wurde abgelehnt. Bitte f&uuml;hren Sie den Bestellvorgang erneut durch.</p>';
+$_['text_payment_payzen_fatal_error'] = '<p>Ein Fehler ist während dem Zahlungsvorgang unterlaufen.</p>';
+$_['text_payment_payzen_payment_error'] = '<p>Ihre Zahlung wurde abgelehnt. Bitte führen Sie den Bestellvorgang erneut durch.</p>';
 
 $_['entry_payment_payzen_trans_id'] = 'Transaktionsnummer';
-$_['entry_payment_payzen_paid_amount'] = 'Paid amount';
+$_['entry_payment_payzen_paid_amount'] = 'Betrag';