+++ upload/catalog/language/es-es/extension/payment/payzen.php
@@ -1,0 +1,26 @@
+<?php
+/**
+ * Copyright © Lyra Network.
+ * This file is part of PayZen plugin for OpenCart See COPYING.md for license details.
+ *
+ * @author     Lyra Network <https://www.lyra.com>
+ * @copyright  Lyra Network
+ * @license    http://www.gnu.org/licenses/gpl.html GNU General Public License (GPL v3)
+ */
+
+$_['text_payment_payzen_title'] = 'Pago con tarjeta de crédito';
+$_['text_payment_payzen_multi_title'] = 'Pago con tarjeta de crédito en cuotas';
+
+$_['heading_title'] = 'El pago fue exitoso. Su pedido se registró correctamente!';
+$_['heading_error_title'] = 'Su pago no fue aceptado!';
+
+$_['text_payment_payzen_maint_warn'] = 'La tienda está en modo de mantenimiento. La notificación automática no puede funcionar.';
+$_['text_payment_payzen_url_check_warn'] = 'La validación automática no ha funcionado. ¿Configuró correctamente la URL de notificación en su Back Office PayZen?';
+$_['text_payment_payzen_url_check_details'] = 'Para entender el problema, lea la documentación del módulo: <br />Capítulo « Leer detenidamente antes de continuar »<br />Capítulo « Configuración de la URL de notificación »';
+$_['text_payment_payzen_pass_to_prod_info'] = '<p><u>IR A PRODUCTION</u></p>Si desea saber cómo poner su tienda en modo production, lea los capítulos « Proceder a la fase de prueba » y « Paso de una tienda al modo producción » en la documentación del módulo.';
+
+$_['text_payment_payzen_fatal_error'] = '<p>Ocurrió un error durante el proceso de pago.</p>';
+$_['text_payment_payzen_payment_error'] = '<p>Su pago no fue aceptado. Intente realizar de nuevo el pedido.</p>';
+
+$_['entry_payment_payzen_trans_id'] = 'ID de transacción';
+$_['entry_payment_payzen_paid_amount'] = 'Monto';