+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+4.1.0, 2020-08-12:
+- Bug fix: payment in installments method was displayed in the frontend even if the multi feature is disabled.
+- Added Spanish translation.
+- Update plugin translations (English, German and French).
+- [prodfaq] Fix notice about shifting the shop to production mode.
+
 4.0.0, 2018-07-16:
 - New PayZen payment module for OpenCart 3.x versions.